refactor(): extract the replacer module
